# swagger.yaml
openapi: 3.0.0
info:
  title: GitGood Invoice Processing API
  description: >
    A cloud-based REST API for generating, converting, validating, finalising,
    and downloading UBL 2.1 compliant electronic invoices.
  version: 1.0.0

servers:
  - url: https://gitgood.onrender.com/
    description: Production (Render)

security:
  - ApiKeyAuth: []

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key

  schemas:
    Item:
      type: object
      required: [itemName, quantity, unitPrice, unitCode, totalPrice]
      properties:
        itemName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        unitCode:
          type: string
          description: "(example: kg)"
        totalPrice:
          type: number

    PaymentDetails:
      type: object
      required: [bankName, accountNumber, bsbAbnNumber, paymentMethod]
      properties:
        bankName:
          type: string
        accountNumber:
          type: string
        bsbAbnNumber:
          type: string
        paymentMethod:
          type: string
          description: "(example: bankTransfer)"

    Error:
      type: object
      properties:
        error:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Unable to complete"

paths:
  /v1/health:
    get:
      summary: Health check
      description: Returns the current status of the service. No authentication required.
      security: []
      tags: [Health]
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  timestamp:
                    type: string
                    format: date-time
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: down
                  reason:
                    type: string
                    example: "Database connection failed"

  /v1/invoice:
    post:
      summary: Create a new invoice
      description: Creates a new invoice from manually entered data in draft status.
      tags: [Invoice Creation]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - buyerName
                - buyerAbn
                - supplierName
                - supplierAbn
                - issueDate
                - paymentDueDate
                - itemsList
                - taxRate
                - paymentDetails
              properties:
                buyerName:
                  type: string
                buyerAbn:
                  type: string
                supplierName:
                  type: string
                supplierAbn:
                  type: string
                issueDate:
                  type: string
                  format: date
                paymentDueDate:
                  type: string
                  format: date
                itemsList:
                  type: array
                  items:
                    $ref: '#/components/schemas/Item'
                taxRate:
                  type: number
                  description: "(example: 0.1)"
                paymentDetails:
                  type: array
                  items:
                    $ref: '#/components/schemas/PaymentDetails'
                additionalNotes:
                  type: string
      responses:
        '201':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  status:
                    type: string
                    example: draft
                  createdAt:
                    type: string
                    format: date-time
        '400':
          description: Missing or invalid fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVALID_REQUEST"
                  message: Missing or Invalid Fields
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INTERNAL_SERVER_ERROR"
                  message: Unexpected error on the server side.

    get:
      summary: List all invoices
      description: Returns list of invoices with optional date filters.
      tags: [Invoice Management]
      parameters:
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limitPerPage
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of Invoices generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoices:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
        '400':
          description: Invalid filter values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVALID_REQUEST"
                  message: Missing or Invalid Fields

  /v1/invoice/{invoiceId}:
    get:
      summary: Get a specific invoice
      tags: [Invoice Management]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice details
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  status:
                    type: string
                    example: draft
                  buyerName:
                    type: string
                  supplierName:
                    type: string
                  issueDate:
                    type: string
                    format: date
                  paymentDueDate:
                    type: string
                    format: date
                  itemsList:
                    type: array
                    items:
                      type: object
                  taxRate: 
                    type: number
                  taxAmount: 
                    type: number
                  totalPayable:
                    type: number
                  ublXml: 
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt: 
                    type: string
                    format: date-time

        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found
        '400':
          description: Invalid parameter value 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVALID_REQUEST"
                  message: Missing or Invalid Fields


    put:
      summary: Update an invoice 
      description: Update an invoice using its ID
      tags: [Invoice Management]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                buyerName:
                  type: string
                buyerAbn:
                  type: string
                supplierName:
                  type: string
                supplierAbn:
                  type: string
                issueDate:
                  type: string
                  format: date
                paymentDueDate:
                  type: string
                  format: date
                itemsList:
                  type: array
                  items:
                    $ref: '#/components/schemas/Item'
                taxRate:
                  type: number
                paymentDetails:
                  type: array
                  items:
                    $ref: '#/components/schemas/PaymentDetails'
                additionalNotes:
                  type: string
      responses:
        '200':
          description: Invoice updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  status:
                    type: string
                    example: draft
                  updatedAt:
                    type: string
                    format: date
        '400':
          description: Invalid parameter value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVALID_REQUEST"
                  message: Missing or Invalid Fields
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found
        '409':
          description: Invoice already finalised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "ALREADY_FINALISED"
                  message: Invoice has already been fianlised




    delete:
      summary: Delete an invoice
      description: delete an invoice from database
      tags: [Invoice Management]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice deleted Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  message:
                    type: string 
                    example: Invoice successfully deleted
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found

  /v1/invoice/{invoiceId}/convert:
    post:
      summary: Convert invoice to UBL XML
      tags: [Invoice Conversion]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice converted to UBL XML Successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  ublXml:
                    type: string 
                  status:
                    type: string 
                    example: "converted"
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found
        '409':
          description: Invoice already converted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "ALREADY_CONVERTED"
                  message: Invoice has already been converted
        '422':
          description: Insufficient data for conversion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INSUFFICIENT_DATA"
                  message: Data is Insufficient for conversion

  /v1/invoice/{invoiceId}/validate:
    post:
      summary: Validate an invoice
      description: Validate an invoice converted to UBL XML format
      tags: [Invoice Validation]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  valid:
                    type: boolean 
                  errors:
                    type: array
                    items:
                      type: object
                    description: "empty if valid"
                  status:
                    type: string
                    example: validated

        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found
        '409':
          description: Invoice not yet converted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVOICE_NOT_CONVERTED"
                  message: Invoice not yet converted

  /v1/invoice/{invoiceId}/final:
    post:
      summary: Finalise a validated invoice
      description: Finalise invoice to be ready to download/email/etc
      tags: [Invoice Finalisation]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice finalised
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceId:
                    type: string
                  ublXml:
                    type: string 
                  status:
                    type: string 
                    example: "finalised"
                  finalisedAt:
                    type: string
                    format: date
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found
        '409':
          description: Invoice not yet validated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVOICE_NOT_VALIDATED"
                  message: Invoice not yet validated

  /v1/invoice/{invoiceId}/download:
    get:
      summary: Download invoice
      description: Download a finalised invoice
      tags: [Invoice Download]
      parameters:
        - in: path
          name: invoiceId
          required: true
          schema:
            type: string
        - in: query
          name: format
          schema:
            type: string
            enum: [xml, json]
            default: xml
      responses:
        '200':
          description: Invoice file downloaded Successfully
        '400':
          description: Invalid format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVALID_REQUEST"
                  message: Missing or Invalid Fields
        '404':
          description: Invoice not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "NOT_FOUND"
                  message: Invoice not found

        '409':
          description: Invoice not yet finalised
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                  error: "INVOICE_NOT_READY"
                  message: Invoice not yet converted/validated